+++ var/ci/phpstan/phpstan-bootstrap.php
@@ -1,0 +1,10 @@
+<?php
+
+/**
+ * Copyright (c) 2017-2020 gyselroth™  (http://www.gyselroth.net)
+ *
+ * @package \gyselroth\Helper
+ * @author  gyselroth™  (http://www.gyselroth.com)
+ * @link    http://www.gyselroth.com
+ * @license Apache-2.0
+ */