+++ CHANGELOG.md
@@ -5,7 +5,8 @@ Version 0.2.3 - Not released yet
 --------------------------------
 * Correct formatting to follow PSR-2 more closely
 * Performance: Root-namespace all PSL functions 
-* Declare more data types in method signatures 
+* Declare more data types in method signatures
+* Add phpstan config 
 
 
 Version 0.2.2 - Released 2019/12/12